feat(new command): updates new to download multiple file templates
Uses the new layout of the functions-templates project to download multiple file templates into a
twilio-run project. Templates will be downloaded into the functions/{bundleName} and
assets/{bundleName} directories within the project.

This also allows for passing an empty bundleName (or filename in the twilio-run CLI parlance right
 now). When empty, this adds the functions/assets directly into the functions/assets directories.
This will be used for `create-twilio-function` template options. The twilio-run CLI still enforces
a name.

BREAKING CHANGE: This needs the functions-templates repo to be up to date. Currently it points to
the next branch, which is up to date.

re #20