feat: introduce config file functionality (#15) (#38)
* feat: introduce config file functionality (#15)

With this change we are starting to allow .twilio-functions for more than just storing some build outcomes. It supports grouping configs by environment, command or project or nested combinations of these.

Very extended config:

```json
{
	"startConfig": {
		"ngrok": "dom",
		"inspect": ""
	},
	"environments": {
		"dev": {
			"deployConfig": {
				"env": ".env"
			}
		},
		"prod": {
			"deployConfig": {
				"env": ".env.prod"
			}
		}
	},
	"projects": {
		"ACc2bdaa19578061b45a518axxxxxxxxxx": {
			"serviceSid": "ZSd6e3037c5710c8e8c979f5xxxxxxxxxx",
			"latestBuild": "ZBdf824389570e3ee2ebcec7xxxxxxxxxx",
			"deployConfig": {
				"environment": "prod"
			}
		}
	},
	"serviceSid": "ZSd6e3037c5710c8e8c979fxxxxxxxxxx",
	"latestBuild": "ZBdf824389570e3ee2ebcec7xxxxxxxxxx"
}
```

fix: #15

* feat: add config support for list,activate,deploy (#15)

This commit adds config file support for the list, activate and deploy
command.

It also deprecates the --functions-env flag

BREAKING CHANGE: Deprecating --functions-env as an option

fix: #15, fix: #27

* refactor(config): factor out common code for testability

* fix(config): write serviceSid to config

This creates compatibility of config files for twilio-run

* test(config): add basic test for list config

* feat(runtime): handle invalid account sid & new error page

This change will check for valid Account SIDs in local development mode and create an appropriate
error message. It also introduces a new Error response page

BREAKING CHANGE: Error page layout changed

fix #45

* chore(release): 2.0.0-beta.13

* fix: update code for new version of severless-api (#46)

* fix: update code for new version of severless-api

re twilio-labs/serverless-api#8

* chore(npm): upgrade @twilio-labs/serverless-api

* fix(tests): fix integration tests

Remove the reliance on stacktraces in the error integration tests

* chore(release): 2.0.0-rc.0

* chore(npm): upgrade serverless-api library

* chore(release): 2.0.0-rc.1

* feat(new): change from prompts to inquirer (#36)

Prompts has a bug for screens that are smaller than the content. This fix hasn't been launched yet
so we are switching to inquirer.

fixes #36

* fix(templates): switch template list endpoint to next branch

The templates are currently pulled from the next branch but the list was from master, causing some
templates to crash

* feat: introduce config file functionality (#15)

With this change we are starting to allow .twilio-functions for more than just storing some build
outcomes. It supprts grouping configs by environment, command or project or nested combinations of
these

fix: #15

* feat: add config support for list,activate,deploy (#15)

This commit adds config file support for the list, activate and deploy
command.

It also deprecates the --functions-env flag

BREAKING CHANGE: Deprecating --functions-env as an option

fix: #15, fix: #27

* refactor(config): factor out common code for testability

* fix(config): write serviceSid to config

This creates compatibility of config files for twilio-run

* test(config): add basic test for list config

* feat(config): accept external config options

* test(config): add more tests for configs

* fix(config): store under accountSid not API key