feat: add estimated execution time to function run
- use process.hrtime.bigint() to track start/stop
- debug log estimated execution time in milliseconds

Not sure if requiring debug() is the correct way to report this information.  We could also consider displaying a warning/error message if the estimated duration execeeds the defined limit.

Another approach to capturing timing would be to use perf_hooks to setup an observer and add measurements.  The observer could process the measurements (logging) and disconnect.