fix(deploy): stop creating environments with domain suffix 'undefined' (#50)
* fix(deploy): stop creating environments with domain suffix 'undefined'

According to https://github.com/twilio-labs/plugin-serverless/issues/35 deploying using the  flag lead to a URL with 'undefined' in the domain. I think this used to work, but there was either a change in the behaviour of  around sending undefined values in form data between version 9.6.0 and the current version 11.0.1, or the Serverless API changed. I started thinking it was the API, but I'm closer to believing it was got now.

Anyway, not sending the  key when it is an empty string instead of sending  should now work.

* test(environments): adds tests for `createEnvironmentForSuffix`