fix(debugger): fixes debugger redaction in plugin-serverless
Fixes #185.

When using the twilio-run debugger within the serverless plugin, redaction wasn't working. I thought this might be because the debug module was being overwritten. Looking at where the formatters are added, the resultant debug object was never exported and used.

Exporting this debugger and explicitly using it brought the formatter back in the plugin.