fix(serverless-api/logs): keeps the most recent logs in log cache
Another note, I discovered while testing this that _read() is called multiple times and lead to extra API calls via _poll(). This was because when calling _read() with an interval set, the code would make a one off call. Now, if an interval is set already, then no extra polls will be made.