feat: adds twilio to package.json in create-twilio-function
When installing twilio-run as a devDependency of a local function repo you get the latest version of twilio installed. BDeploying that function will then use the default version of twilio in Functions which is falling further and further behind. This update adds twilio as a direct dependency and installs the latest version allowing developers to see the dependency, understand the version and use their functions locally and deployed with confidence.