fix(serverless-api): fixes assets paths in deployment (#258)
So, as part of the assets plugin, I decided that encoding the path of an asset was a good thing and I decided to do so at the point it is sent to the API in the serverless-api project. However, I forgot that that would affect assets getting deployed from twilio-run. Asset paths were getting their initial forward slash encoded and that broke their expected URLs/paths.

I have reverted the change in serverless API and just URI encode the path from inside the assets plugin now.