fix(twilio-run): show service sid if available instead of name
If a Service SID is defined in your config or .twiliodeployinfo we should pick that to display
instead of the name as the name might not be the actual name of the service. Similarly in the deploy
result we'll actually show the real service name not the presumed one.

BREAKING CHANGE: The output of the config info in deployment changed

fix #262