fix(twilio-run): fix global scope in forked process
We were requiring the respective Function file before the global scope was created. If a user was
referencing a global object such as `Runtime` outside of the `handler` function it would throw an
UnhandledPromiseException because the Runtime wasn't constructe yet. This change only requires the
Function right before execution. This will also create more acurate timing results for cold starts.
During the fix I realized that the Function was also required twice, once by the Server and once by
the forked process. The server should only require the Function if it's not a forked process.