fix(plugin-serverless): remove default from aliases
The recent introduction of alias support in the `plugin-serverless` was causing issues if there was
a default value as the default value of the alias would override the value explicitly passed into
the original command. This change removes default and required arguments from aliases and also
explicitly updates the description.

fix #272