refactor(create-twilio-function): handle streams with callbacks
Within the createGitignore function we use fs.promises and also createWriteStream and the two aren't particularly compatible. createWriteStream far prefers callbacks and events.

This PR rewrites createGitignore to use callbacks and streams internally returning a Promise for external users.

This also adds the .twiliodeployinfo line to the top of the gitignore file rather than the bottom. This avoids having to open the file twice (gitignore.writeFile will end the write stream when it pipes all the content to it). I think this is the best place to put the addition as it will be where a reader starts looking when they open the gitignore file too.