chore(twilio-run): fix types for printRouteInfo
I didn't feel comfortable making pkgJson optional on StartCliConfig, but that was initially how I
fixed a type issue in printRouteInfo. The real issue was that printRouteInfo was being used with
both StartCliConfig and ServerConfig typed objects. It was better to call out that this is how
the function was being used, then handle the different types within. This uses a hasOwnProperty
utility that properly narrows down the type on an object based on the available properties, thanks
to Stefan Baumgartner and allows the existing code to continue working as expected.