fix(runtime-handler): using set-cookie now sets cookie header
Following the implementation of the Functions runtime-handler, the Response object
sets the Set-Cookie header to [] if there are no cookies set. If you then set a
header with the key 'set-cookie' (or any other capitalisation) it will be eventually
overridden by the empty array under Set-Cookie (since it is also added to the headers last).
To fix this all cookies are set under the key 'Set-Cookie'.