chore(deps): brings @types/jest up to date
Updating things in serverless-api caused the integration test to stop
loading because 'os.networkInterfaces isn't a function'. This had
something to do with mocking the os module, which I fixed with the
addition to the integration test. I don't know why it wasn't broken
before.

This makes the version of @types/jest the same across the packages and
matches the version of jest across the packages too.

Supersedes #351