chore(serverless-api): fixes integration test mock
This still wasn't working. The module escaped being mocked and failed in
CI on non-macOS systems. Updated the mock to use mock values, not a
possible real value. Then updated the tests to ensure that the mock
values were going through the code. Finally worked out how to correctly
require the mock os and extend it for the integration test to stop the
entire test failing to compile due to missing os properties when they
weren't mocked.