fix(create-twilio-function): adds rootDir to generated tsconfig.json
By default, tsc deems the lowest common directory as the root directory
of the application and delivers the compiled files into the outDir.
If you do not have an `assets` directory with a .ts file in it, then tsc
deems src/functions as the root directory and puts the compiled
functions directly into the `dist` directory. This wasn't obvious,
because the example generated files included a private TS asset.

Setting the `rootDir` to "src" ensures that the base directory is "src"
and a "functions" directory is generated in `dist`.