chore: update dependencies (#359)
* chore(twilio-run): updates boxen

Also, boxen is now written in TypeScript and has .d.ts files, so removed boxen from types.d.ts

* chore(create-twilio-function): updates boxen dependency

* chore(deps): brings @types/jest up to date

Updating things in serverless-api caused the integration test to stop
loading because 'os.networkInterfaces isn't a function'. This had
something to do with mocking the os module, which I fixed with the
addition to the integration test. I don't know why it wasn't broken
before.

This makes the version of @types/jest the same across the packages and
matches the version of jest across the packages too.

Supersedes #351

* chore(plugin-serverless): update dependencies

Updates @oclif/plugin-help and @twilio/cli-core. Supersedes #358

* chore(twilio-run): updates wrap-ansi

* chore(twilio-run): updates yargs & inquirer

* chore(create-twilio-function): updates yargs, inquirer, wrap-ansi

* chore(twilio-run): update ora

* chore(create-twilio-function): update ora

* chore(serverless-api): fixes integration test mock

This still wasn't working. The module escaped being mocked and failed in
CI on non-macOS systems. Updated the mock to use mock values, not a
possible real value. Then updated the tests to ensure that the mock
values were going through the code. Finally worked out how to correctly
require the mock os and extend it for the integration test to stop the
entire test failing to compile due to missing os properties when they
weren't mocked.

Co-authored-by: Dominik Kundel <dkundel@twilio.com>