fix(create-twilio-function): handle rimraf failing
Received a report that creating a new project had failed for a developer
and then the cleanup task didn't end. There was an issue with deleting
the files that had been created which threw an unhandled error in the
cleanUpAndExit function leading to the spinner never stopping.

This adds a try/catch/finally around rimraf in cleanUpAndExit to ensure
that it does indeed exit.