chore(deps): upgrades typescript to 4.2.4
Fixes one type issue that came up in the upgrade. The project builds
now.

4.2.4 is not the latest version of TS for now. Reports around the
internet showed that upgrading to TS 4.2 fixed the Prettier issues we've
been seeing. Upgrading all the way to 4.7.x caused a bunch more type
issues. We should plan an upgrade, in order to keep up, but there might
be further effects.