chore(runtime-handler): update default twilio to 3.80.0
This removes the vulnerability alerts for @twilio/runtime-handler but
might cause some side-effects for the edge case where a customer has
specified this version of runtime-handler but has not specified a
version of twilio. In this case the local and deployed behavior
might vary depending on the APIs used by the customer in Functions.